feat: rename PK-Concatenator file to PartitionKeyConcatenator
